Moving SCOPED_AUTH.
