+++ packages/mds-api-scopes/tsconfig.build.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "outDir": "dist"
   },
-  "references": []
+  "references": [{ "path": "../../packages/mds-types/tsconfig.build.json" }]
 }
