+++ packages/mds-api-server/index.ts
@@ -2,7 +2,8 @@ import bodyParser from 'body-parser'
 import express from 'express'
 import { pathsFor, AuthorizationError } from '@mds-core/mds-utils'
 import { AuthorizationHeaderApiAuthorizer, ApiAuthorizer, ApiAuthorizerClaims } from '@mds-core/mds-api-authorizer'
-import { ScopeValidator, validateScopes, AccessTokenScope } from '@mds-core/mds-api-scopes'
+import { validateScopes } from '@mds-core/mds-api-scopes'
+import { ScopeValidator, AccessTokenScope } from '@mds-core/mds-types'
 
 export type ApiRequest = express.Request
