+++ packages/mds-audit/tests/api.spec.ts
@@ -31,14 +31,19 @@ import {
   PROPULSION_TYPES,
   VEHICLE_TYPES
 } from '@mds-core/mds-types'
-import { PROVIDER_UUID, makeEventsWithTelemetry, makeDevices, makeTelemetryInArea } from '@mds-core/mds-test-data'
+import {
+  PROVIDER_UUID,
+  makeEventsWithTelemetry,
+  makeDevices,
+  makeTelemetryInArea,
+  SCOPED_AUTH
+} from '@mds-core/mds-test-data'
 import { now, rangeRandomInt } from '@mds-core/mds-utils'
 import cache from '@mds-core/mds-cache'
 import test from 'unit.js'
 import uuid from 'uuid'
 import { ApiServer } from '@mds-core/mds-api-server'
 import db from '@mds-core/mds-db'
-import { AccessTokenScope } from '@mds-core/mds-api-scopes'
 import { api } from '../api'
 
 const request = supertest(ApiServer(api))
@@ -65,9 +70,6 @@ const telemetry = (): {} => ({
 const AUDIT_START = Date.now()
 
 const audit_subject_id = 'user@mds-testing.info'
-
-const SCOPED_AUTH = (...scopes: AccessTokenScope[]) =>
-  `basic ${Buffer.from(`${audit_subject_id}|${scopes.join(' ')}`).toString('base64')}`
 
 before('Initializing Database', async () => {
   await Promise.all([db.initialize(), cache.initialize()])
@@ -105,7 +107,7 @@ describe('Testing API', () => {
   it('verify audit start (matching vehicle)', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/start`)
-      .set('Authorization', SCOPED_AUTH('audits:write'))
+      .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
       .send({
         audit_event_id: uuid(),
         timestamp: AUDIT_START,
@@ -128,7 +130,7 @@ describe('Testing API', () => {
   it('verify audit start (conflict)', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/start`)
-      .set('Authorization', SCOPED_AUTH('audits:write'))
+      .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
       .expect(409)
       .end((err, result) => {
         test.value(result).hasHeader('content-type', APP_JSON)
@@ -139,7 +141,7 @@ describe('Testing API', () => {
   it('verify audit start (no scope)', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/start`)
-      .set('Authorization', SCOPED_AUTH())
+      .set('Authorization', SCOPED_AUTH([], ''))
       .expect(403)
       .end((err, result) => {
         test.value(result).hasHeader('content-type', APP_JSON)
@@ -150,7 +152,7 @@ describe('Testing API', () => {
   it('verify audit issue event', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/event`)
-      .set('Authorization', SCOPED_AUTH('audits:write'))
+      .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
       .send({
         audit_event_id: uuid(),
         audit_event_type: AUDIT_EVENT_TYPES.issue,
@@ -170,7 +172,7 @@ describe('Testing API', () => {
   it('verify audit issue event (no scope)', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/event`)
-      .set('Authorization', SCOPED_AUTH())
+      .set('Authorization', SCOPED_AUTH([], ''))
       .send({
         audit_event_id: uuid(),
         audit_event_type: AUDIT_EVENT_TYPES.issue,
@@ -189,7 +191,7 @@ describe('Testing API', () => {
   it('verify trip start event', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/vehicle/event`)
-      .set('Authorization', SCOPED_AUTH('audits:write'))
+      .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
       .send({
         event_type: VEHICLE_EVENTS.trip_start,
         timestamp: Date.now(),
@@ -207,7 +209,7 @@ describe('Testing API', () => {
   it('verify audit telemetry', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/vehicle/telemetry`)
-      .set('Authorization', SCOPED_AUTH('audits:write'))
+      .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
       .send({
         telemetry: telemetry(),
         timestamp: Date.now()
@@ -223,7 +225,7 @@ describe('Testing API', () => {
   it('verify audit telemetry (no scope)', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/vehicle/telemetry`)
-      .set('Authorization', SCOPED_AUTH())
+      .set('Authorization', SCOPED_AUTH([], ''))
       .send({
         telemetry: telemetry(),
         timestamp: Date.now()
@@ -238,7 +240,7 @@ describe('Testing API', () => {
   it('verify trip end event', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/vehicle/event`)
-      .set('Authorization', SCOPED_AUTH('audits:write'))
+      .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
       .send({
         event_type: VEHICLE_EVENTS.trip_end,
         timestamp: Date.now(),
@@ -256,7 +258,7 @@ describe('Testing API', () => {
   it('verify audit summary event', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/event`)
-      .set('Authorization', SCOPED_AUTH('audits:write'))
+      .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
       .send({
         audit_event_id: uuid(),
         audit_event_type: AUDIT_EVENT_TYPES.summary,
@@ -275,7 +277,7 @@ describe('Testing API', () => {
   it('verify audit end', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/end`)
-      .set('Authorization', SCOPED_AUTH('audits:write'))
+      .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
       .send({
         audit_event_id: uuid(),
         timestamp: Date.now(),
@@ -292,7 +294,7 @@ describe('Testing API', () => {
   it('verify audit end (no scope)', done => {
     request
       .post(`/audit/trips/${audit_trip_id}/end`)
-      .set('Authorization', SCOPED_AUTH())
+      .set('Authorization', SCOPED_AUTH([], ''))
       .send({
         audit_event_id: uuid(),
         timestamp: Date.now(),
@@ -308,7 +310,7 @@ describe('Testing API', () => {
   it('verify get audit (matched vehicle)', done => {
     request
       .get(`/audit/trips/${audit_trip_id}`)
-      .set('Authorization', SCOPED_AUTH('audits:read'))
+      .set('Authorization', SCOPED_AUTH(['audits:read'], audit_subject_id))
       .expect(200)
       .end((err, result) => {
         test.value(result).hasHeader('content-type', APP_JSON)
@@ -323,7 +325,7 @@ describe('Testing API', () => {
     it(`verify post audit (not found) ${route}`, done => {
       request
         .post(route)
-        .set('Authorization', SCOPED_AUTH('audits:write'))
+        .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
         .expect(404)
         .end((err, result) => {
           test.value(result).hasHeader('content-type', APP_JSON)
@@ -335,7 +337,7 @@ describe('Testing API', () => {
   it(`verify get audit (no scope)`, done => {
     request
       .get(`/audit/trips/${uuid()}`)
-      .set('Authorization', SCOPED_AUTH())
+      .set('Authorization', SCOPED_AUTH([], ''))
       .expect(403)
       .end((err, result) => {
         test.value(result).hasHeader('content-type', APP_JSON)
@@ -346,7 +348,7 @@ describe('Testing API', () => {
   it(`verify get audit (not found)`, done => {
     request
       .get(`/audit/trips/${uuid()}`)
-      .set('Authorization', SCOPED_AUTH('audits:read'))
+      .set('Authorization', SCOPED_AUTH(['audits:read'], audit_subject_id))
       .expect(404)
       .end((err, result) => {
         test.value(result).hasHeader('content-type', APP_JSON)
@@ -357,7 +359,7 @@ describe('Testing API', () => {
   it(`verify list audits (no scope)`, done => {
     request
       .get(`/audit/trips`)
-      .set('Authorization', SCOPED_AUTH())
+      .set('Authorization', SCOPED_AUTH([], ''))
       .expect(403)
       .end((err, result) => {
         test.value(result).hasHeader('content-type', APP_JSON)
@@ -383,7 +385,7 @@ describe('Testing API', () => {
     it(`verify list audits (Filter: ${filter}, Count: ${count})`, done => {
       request
         .get(`/audit/trips${query}`)
-        .set('Authorization', SCOPED_AUTH('audits:read'))
+        .set('Authorization', SCOPED_AUTH(['audits:read'], audit_subject_id))
         .expect(200)
         .end((err, result) => {
           test.value(result).hasHeader('content-type', APP_JSON)
@@ -397,7 +399,7 @@ describe('Testing API', () => {
   it('verify delete audit (no scope)', done => {
     request
       .delete(`/audit/trips/${uuid()}`)
-      .set('Authorization', SCOPED_AUTH())
+      .set('Authorization', SCOPED_AUTH([], ''))
       .expect(403)
       .end((err, result) => {
         test.value(result).hasHeader('content-type', APP_JSON)
@@ -408,7 +410,7 @@ describe('Testing API', () => {
   it('verify delete audit (not found)', done => {
     request
       .delete(`/audit/trips/${uuid()}`)
-      .set('Authorization', SCOPED_AUTH('audits:delete'))
+      .set('Authorization', SCOPED_AUTH(['audits:delete'], audit_subject_id))
       .expect(404)
       .end((err, result) => {
         test.value(result).hasHeader('content-type', APP_JSON)
@@ -419,7 +421,7 @@ describe('Testing API', () => {
   it('verify delete audit', done => {
     request
       .delete(`/audit/trips/${audit_trip_id}`)
-      .set('Authorization', SCOPED_AUTH('audits:delete'))
+      .set('Authorization', SCOPED_AUTH(['audits:delete'], audit_subject_id))
       .expect(200)
       .end((err, result) => {
         test.value(result).hasHeader('content-type', APP_JSON)
@@ -431,7 +433,7 @@ describe('Testing API', () => {
   it('verify audit start (missing vehicle)', done => {
     request
       .post(`/audit/trips/${audit_trip_id_2}/start`)
-      .set('Authorization', SCOPED_AUTH('audits:write'))
+      .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
       .send({
         audit_event_id: uuid(),
         timestamp: AUDIT_START,
@@ -453,7 +455,7 @@ describe('Testing API', () => {
   it('verify get audit (missing vehicle)', done => {
     request
       .get(`/audit/trips/${audit_trip_id_2}`)
-      .set('Authorization', SCOPED_AUTH('audits:read'))
+      .set('Authorization', SCOPED_AUTH(['audits:read'], audit_subject_id))
       .expect(200)
       .end((err, result) => {
         test.value(result).hasHeader('content-type', APP_JSON)
@@ -478,7 +480,7 @@ describe('Testing API', () => {
     it(`verify vehicle matching ${vehicle_id}`, done => {
       request
         .post(`/audit/trips/${uuid()}/start`)
-        .set('Authorization', SCOPED_AUTH('audits:write'))
+        .set('Authorization', SCOPED_AUTH(['audits:write'], audit_subject_id))
         .send({
           audit_event_id: uuid(),
           timestamp: AUDIT_START,
@@ -525,7 +527,7 @@ describe('Testing API', () => {
       const bbox = [[-118.484776, 33.996855], [-118.452283, 33.96299]] // BBOX encompasses the entirity of CANALS
       request
         .get(`/vehicles?bbox=${JSON.stringify(bbox)}`)
-        .set('Authorization', SCOPED_AUTH('audits:vehicles:read'))
+        .set('Authorization', SCOPED_AUTH(['audits:vehicles:read'], audit_subject_id))
         .expect(200)
         .end((err, result) => {
           test.value(result).hasHeader('content-type', APP_JSON)
@@ -543,7 +545,7 @@ describe('Testing API', () => {
       const bbox = [[-118.484776, 33.996855], [-118.452283, 33.96299]] // BBOX encompasses the entirity of CANALS
       request
         .get(`/vehicles?bbox=${JSON.stringify(bbox)}&skip=2&take=5`)
-        .set('Authorization', SCOPED_AUTH('audits:vehicles:read'))
+        .set('Authorization', SCOPED_AUTH(['audits:vehicles:read'], audit_subject_id))
         .expect(200)
         .end((err, result) => {
           test.value(result).hasHeader('content-type', APP_JSON)
@@ -562,7 +564,7 @@ describe('Testing API', () => {
       const bbox = [[-118.484776, 33.996855], [-118.452283, 33.96299]] // BBOX encompasses the entirity of CANALS
       request
         .get(`/vehicles?bbox=${JSON.stringify(bbox)}&skip=10&take=5`)
-        .set('Authorization', SCOPED_AUTH('audits:vehicles:read'))
+        .set('Authorization', SCOPED_AUTH(['audits:vehicles:read'], audit_subject_id))
         .expect(200)
         .end((err, result) => {
           test.value(result).hasHeader('content-type', APP_JSON)
