+++ packages/mds-policy-author/tests/api.spec.ts
@@ -30,8 +30,6 @@ import db from '@mds-core/mds-db'
 import { clone } from '@mds-core/mds-utils'
 import { Policy } from '@mds-core/mds-types'
 import { ApiServer } from '@mds-core/mds-api-server'
-import { AccessTokenScope } from '@mds-core/mds-api-scopes'
-import { TEST1_PROVIDER_ID } from '@mds-core/mds-providers'
 
 import {
   POLICY_JSON,
@@ -44,7 +42,8 @@ import {
   GEOGRAPHY_UUID,
   GEOGRAPHY2_UUID,
   LA_CITY_BOUNDARY,
-  DISTRICT_SEVEN
+  DISTRICT_SEVEN,
+  SCOPED_AUTH
 } from '@mds-core/mds-test-data'
 import { api } from '../api'
 
@@ -54,9 +53,12 @@ const log = console.log.bind(console)
 const request = supertest(ApiServer(api))
 
 const APP_JSON = 'application/json; charset=utf-8'
-
-const SCOPED_AUTH = (...scopes: AccessTokenScope[]) =>
-  `basic ${Buffer.from(`${TEST1_PROVIDER_ID}|${scopes.join(' ')}`).toString('base64')}`
+const EMPTY_SCOPE = SCOPED_AUTH([], '')
+const EVENTS_READ_SCOPE = SCOPED_AUTH(['events:read'])
+const POLICIES_WRITE_SCOPE = SCOPED_AUTH(['policies:write'])
+const POLICIES_READ_SCOPE = SCOPED_AUTH(['policies:read'])
+const POLICIES_PUBLISH_SCOPE = SCOPED_AUTH(['policies:publish'])
+const POLICIES_DELETE_SCOPE = SCOPED_AUTH(['policies:delete'])
 
 describe('Tests app', () => {
   describe('Policy tests', () => {
@@ -72,7 +74,7 @@ describe('Tests app', () => {
       const policy = POLICY_JSON
       request
         .post(`/policies`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .send(policy)
         .expect(403)
         .end(err => {
@@ -84,7 +86,7 @@ describe('Tests app', () => {
       const policy = POLICY_JSON
       request
         .post(`/policies`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .send(policy)
         .expect(403)
         .end(err => {
@@ -98,7 +100,7 @@ describe('Tests app', () => {
       const bad_policy = bad_policy_json
       request
         .post(`/policies`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(bad_policy)
         .expect(400)
         .end((err, result) => {
@@ -113,7 +115,7 @@ describe('Tests app', () => {
       const policy = clone(POLICY_JSON)
       request
         .put(`/policies/d2e31798-f22f-4034-ad36-1f88621b276a`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .send(policy)
         .expect(403)
         .end((err, result) => {
@@ -126,7 +128,7 @@ describe('Tests app', () => {
       const policy = clone(POLICY_JSON)
       request
         .put(`/policies/d2e31798-f22f-4034-ad36-1f88621b276a`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .send(policy)
         .expect(403)
         .end((err, result) => {
@@ -139,7 +141,7 @@ describe('Tests app', () => {
       const policy = clone(POLICY_JSON)
       request
         .put(`/policies/d2e31798-f22f-4034-ad36-1f88621b276a`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(policy)
         .expect(404)
         .end((err, result) => {
@@ -152,7 +154,7 @@ describe('Tests app', () => {
       const policy = POLICY_JSON
       request
         .post(`/policies`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(policy)
         .expect(201)
         .end((err, result) => {
@@ -165,7 +167,7 @@ describe('Tests app', () => {
       const policy = POLICY_JSON
       request
         .post(`/policies`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(policy)
         .expect(409)
         .end((err, result) => {
@@ -180,7 +182,7 @@ describe('Tests app', () => {
       const bad_policy = bad_policy_json
       await request
         .put(`/policies/${POLICY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(bad_policy)
         .expect(400)
     })
@@ -190,7 +192,7 @@ describe('Tests app', () => {
       policy.name = 'a shiny new name'
       await request
         .put(`/policies/${POLICY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(policy)
         .expect(200)
 
@@ -202,7 +204,7 @@ describe('Tests app', () => {
       const policy2 = POLICY2_JSON
       request
         .post(`/policies`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(policy2)
         .expect(201)
         .end((err, result) => {
@@ -216,7 +218,7 @@ describe('Tests app', () => {
       const policy3 = POLICY3_JSON
       request
         .post(`/policies`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(policy3)
         .expect(201)
         .end((err, result) => {
@@ -228,7 +230,7 @@ describe('Tests app', () => {
     it('verifies cannot publish a policy (no auth)', done => {
       request
         .post(`/policies/${POLICY_JSON.policy_id}/publish`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .expect(403)
         .end(err => {
           done(err)
@@ -238,7 +240,7 @@ describe('Tests app', () => {
     it('verifies cannot publish a policy (wrong auth)', done => {
       request
         .post(`/policies/${POLICY_JSON.policy_id}/publish`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .expect(403)
         .end(err => {
           done(err)
@@ -248,7 +250,7 @@ describe('Tests app', () => {
     it('verifies cannot publish a policy with missing geographies', done => {
       request
         .post(`/policies/${POLICY_JSON.policy_id}/publish`)
-        .set('Authorization', SCOPED_AUTH('policies:publish'))
+        .set('Authorization', POLICIES_PUBLISH_SCOPE)
         .expect(404)
         .end((err, result) => {
           test.value(result).hasHeader('content-type', APP_JSON)
@@ -259,7 +261,7 @@ describe('Tests app', () => {
     it('verifies cannot publish a policy that was never POSTed', done => {
       request
         .post(`/policies/${GEOGRAPHY_UUID}/publish`)
-        .set('Authorization', SCOPED_AUTH('policies:publish'))
+        .set('Authorization', POLICIES_PUBLISH_SCOPE)
         .expect(404)
         .end((err, result) => {
           test.value(result).hasHeader('content-type', APP_JSON)
@@ -271,7 +273,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: LA_CITY_BOUNDARY }
       request
         .post(`/geographies`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(geography)
         .expect(201)
         .end((err, result) => {
@@ -285,7 +287,7 @@ describe('Tests app', () => {
     it('can publish a policy', done => {
       request
         .post(`/policies/${POLICY_JSON.policy_id}/publish`)
-        .set('Authorization', SCOPED_AUTH('policies:publish'))
+        .set('Authorization', POLICIES_PUBLISH_SCOPE)
         .expect(200)
         .end((err, result) => {
           test.value(result).hasHeader('content-type', APP_JSON)
@@ -296,7 +298,7 @@ describe('Tests app', () => {
     it('cannot double-publish a policy', done => {
       request
         .post(`/policies/${POLICY_JSON.policy_id}/publish`)
-        .set('Authorization', SCOPED_AUTH('policies:publish'))
+        .set('Authorization', POLICIES_PUBLISH_SCOPE)
         .expect(409)
         .end((err, result) => {
           test.value(result).hasHeader('content-type', APP_JSON)
@@ -309,7 +311,7 @@ describe('Tests app', () => {
       policy.name = 'an even shinier new name'
       request
         .put(`/policies/${POLICY_JSON.policy_id}`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .send(policy)
         .expect(403)
         .end(err => {
@@ -322,7 +324,7 @@ describe('Tests app', () => {
       policy.name = 'an even shinier new name'
       request
         .put(`/policies/${POLICY_JSON.policy_id}`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .send(policy)
         .expect(403)
         .end(err => {
@@ -335,7 +337,7 @@ describe('Tests app', () => {
       policy.name = 'an even shinier new name'
       request
         .put(`/policies/${POLICY_JSON.policy_id}`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(policy)
         .expect(409)
         .end((err, result) => {
@@ -347,7 +349,7 @@ describe('Tests app', () => {
     it('cannot delete an unpublished policy (no auth)', done => {
       request
         .delete(`/policies/${POLICY2_UUID}`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .expect(403)
         .end(async err => {
           done(err)
@@ -357,7 +359,7 @@ describe('Tests app', () => {
     it('cannot delete an unpublished policy (wrong auth)', done => {
       request
         .delete(`/policies/${POLICY2_UUID}`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .expect(403)
         .end(async err => {
           done(err)
@@ -367,7 +369,7 @@ describe('Tests app', () => {
     it('can delete an unpublished policy', done => {
       request
         .delete(`/policies/${POLICY2_UUID}`)
-        .set('Authorization', SCOPED_AUTH('policies:delete'))
+        .set('Authorization', POLICIES_DELETE_SCOPE)
         .expect(200)
         .end(async (err, result) => {
           const body = result.body
@@ -381,7 +383,7 @@ describe('Tests app', () => {
     it('cannot delete a published policy', done => {
       request
         .delete(`/policies/${POLICY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('policies:delete'))
+        .set('Authorization', POLICIES_DELETE_SCOPE)
         .expect(404)
         .end(err => {
           done(err)
@@ -392,7 +394,7 @@ describe('Tests app', () => {
       const metadata = { some_arbitrary_thing: 'boop' }
       await request
         .put(`/policies/${POLICY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .send({ policy_id: POLICY_UUID, policy_metadata: metadata })
         .expect(403)
     })
@@ -401,7 +403,7 @@ describe('Tests app', () => {
       const metadata = { some_arbitrary_thing: 'boop' }
       await request
         .put(`/policies/${POLICY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .send({ policy_id: POLICY_UUID, policy_metadata: metadata })
         .expect(403)
     })
@@ -410,7 +412,7 @@ describe('Tests app', () => {
       const metadata = { some_arbitrary_thing: 'boop' }
       await request
         .put(`/policies/${POLICY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send({ policy_id: POLICY_UUID, policy_metadata: metadata })
         .expect(201)
       const result = await db.readSinglePolicyMetadata(POLICY_UUID)
@@ -421,7 +423,7 @@ describe('Tests app', () => {
       const metadata = { some_arbitrary_thing: 'beep' }
       await request
         .put(`/policies/${POLICY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send({ policy_id: POLICY_UUID, policy_metadata: metadata })
         .expect(200)
       const result = await db.readSinglePolicyMetadata(POLICY_UUID)
@@ -431,7 +433,7 @@ describe('Tests app', () => {
     it('cannot GET policy metadata (no auth)', done => {
       request
         .get(`/policies/${POLICY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .expect(403)
         .end(err => {
           done(err)
@@ -441,7 +443,7 @@ describe('Tests app', () => {
     it('cannot GET policy metadata (wrong auth)', done => {
       request
         .get(`/policies/${POLICY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .expect(403)
         .end(err => {
           done(err)
@@ -451,7 +453,7 @@ describe('Tests app', () => {
     it('verifies GETing policy metadata when given a policy_id', done => {
       request
         .get(`/policies/${POLICY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(200)
         .end((err, result) => {
           test.assert(result.body.policy_metadata.some_arbitrary_thing === 'beep')
@@ -463,7 +465,7 @@ describe('Tests app', () => {
     it('verifies cannot GET non-existent policy metadata', done => {
       request
         .get(`/policies/beepbapboop/meta`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(404)
         .end((err, result) => {
           test.assert(result.body.result === 'not found')
@@ -475,21 +477,21 @@ describe('Tests app', () => {
     it('cannot GET policy metadata (no auth)', async () => {
       await request
         .get(`/policies/meta`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .expect(403)
     })
 
     it('cannot GET policy metadata (wrong auth)', async () => {
       await request
         .get(`/policies/meta`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .expect(403)
     })
 
     it('verifies GETting policy metadata with the same params as for bulk policy reads', async () => {
       const result = await request
         .get(`/policies/meta`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(200)
       test.assert(result.body.length === 1)
       test.value(result).hasHeader('content-type', APP_JSON)
@@ -498,7 +500,7 @@ describe('Tests app', () => {
     it('cannot GET a single policy (no auth)', done => {
       request
         .get(`/policies/${POLICY_UUID}`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .expect(403)
         .end(async err => {
           done(err)
@@ -508,7 +510,7 @@ describe('Tests app', () => {
     it('cannot GET a single policy (wrong auth)', done => {
       request
         .get(`/policies/${POLICY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .expect(403)
         .end(async err => {
           done(err)
@@ -518,7 +520,7 @@ describe('Tests app', () => {
     it('can GET a single policy', done => {
       request
         .get(`/policies/${POLICY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(200)
         .end(async (err, result) => {
           test.assert(result.body.policy_id === POLICY_UUID)
@@ -530,7 +532,7 @@ describe('Tests app', () => {
     it('cannot GET a single nonexistent policy', done => {
       request
         .get(`/policies/544d36c4-29f5-4088-a52f-7c9a64d5874c`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(404)
         .end(async err => {
           done(err)
@@ -556,14 +558,14 @@ describe('Tests app', () => {
     it('cannot GET all active policies (no auth)', async () => {
       await request
         .get(`/policies`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .expect(403)
     })
 
     it('cannot GET all active policies (wrong auth)', async () => {
       await request
         .get(`/policies`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .expect(403)
     })
 
@@ -574,7 +576,7 @@ describe('Tests app', () => {
       await db.publishPolicy(SUPERSEDING_POLICY_JSON.policy_id)
       request
         .get(`/policies`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(200)
         .end(async (policies_err, policies_result) => {
           test.assert(policies_result.body.length === 4)
@@ -585,7 +587,7 @@ describe('Tests app', () => {
     it('can GET all published policies', done => {
       request
         .get(`/policies?get_published=true`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(200)
         .end(async (policies_err, policies_result) => {
           test.assert(policies_result.body.length === 2)
@@ -596,7 +598,7 @@ describe('Tests app', () => {
     it('can GET all unpublished policies', done => {
       request
         .get(`/policies?get_unpublished=true`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(200)
         .end(async (policies_err, policies_result) => {
           test.assert(policies_result.body.length === 2)
@@ -607,7 +609,7 @@ describe('Tests app', () => {
     it('throws an exception if both get_unpublished and get_published are submitted', done => {
       request
         .get(`/policies?get_unpublished=true&get_published=true`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(400)
         .end(async policies_err => {
           done(policies_err)
@@ -628,7 +630,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: LA_CITY_BOUNDARY }
       request
         .post(`/geographies`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .send(geography)
         .expect(403)
         .end(err => {
@@ -640,7 +642,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: LA_CITY_BOUNDARY }
       request
         .post(`/geographies`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .send(geography)
         .expect(403)
         .end(err => {
@@ -652,7 +654,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: LA_CITY_BOUNDARY }
       request
         .post(`/geographies`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(geography)
         .expect(201)
         .end((err, result) => {
@@ -666,7 +668,7 @@ describe('Tests app', () => {
     it('cannot GETs one current geography (no auth)', done => {
       request
         .get(`/geographies/${GEOGRAPHY_UUID}`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .expect(403)
         .end(err => {
           done(err)
@@ -676,7 +678,7 @@ describe('Tests app', () => {
     it('cannot GETs one current geography (wrong auth)', done => {
       request
         .get(`/geographies/${GEOGRAPHY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .expect(403)
         .end(err => {
           done(err)
@@ -686,7 +688,7 @@ describe('Tests app', () => {
     it('GETs one current geography', done => {
       request
         .get(`/geographies/${GEOGRAPHY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(200)
         .end((err, result) => {
           test.assert(result.body.geography_id === GEOGRAPHY_UUID)
@@ -698,7 +700,7 @@ describe('Tests app', () => {
     it('cannot GET a nonexistent geography', done => {
       request
         .get(`/geographies/${POLICY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(404)
         .end(err => {
           done(err)
@@ -709,7 +711,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: DISTRICT_SEVEN }
       request
         .put(`/geographies/${GEOGRAPHY_UUID}`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .send(geography)
         .expect(403)
         .end(err => {
@@ -721,7 +723,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: DISTRICT_SEVEN }
       request
         .put(`/geographies/${GEOGRAPHY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .send(geography)
         .expect(403)
         .end(err => {
@@ -733,7 +735,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: DISTRICT_SEVEN }
       request
         .put(`/geographies/${GEOGRAPHY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(geography)
         .expect(201)
         .end((err, result) => {
@@ -746,7 +748,7 @@ describe('Tests app', () => {
       const metadata = { some_arbitrary_thing: 'boop' }
       await request
         .put(`/geographies/${GEOGRAPHY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .send({ geography_id: GEOGRAPHY_UUID, geography_metadata: metadata })
         .expect(403)
     })
@@ -755,7 +757,7 @@ describe('Tests app', () => {
       const metadata = { some_arbitrary_thing: 'boop' }
       await request
         .put(`/geographies/${GEOGRAPHY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .send({ geography_id: GEOGRAPHY_UUID, geography_metadata: metadata })
         .expect(403)
     })
@@ -764,7 +766,7 @@ describe('Tests app', () => {
       const metadata = { some_arbitrary_thing: 'boop' }
       await request
         .put(`/geographies/${GEOGRAPHY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send({ geography_id: GEOGRAPHY_UUID, geography_metadata: metadata })
         .expect(201)
       const result = await db.readSingleGeographyMetadata(GEOGRAPHY_UUID)
@@ -775,7 +777,7 @@ describe('Tests app', () => {
       const metadata = { some_arbitrary_thing: 'beep' }
       await request
         .put(`/geographies/${GEOGRAPHY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send({ geography_id: GEOGRAPHY_UUID, geography_metadata: metadata })
         .expect(200)
       const result = await db.readSingleGeographyMetadata(GEOGRAPHY_UUID)
@@ -785,7 +787,7 @@ describe('Tests app', () => {
     it('cannot GET geography metadata (no auth)', done => {
       request
         .get(`/geographies/${GEOGRAPHY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .expect(403)
         .end(err => {
           done(err)
@@ -795,7 +797,7 @@ describe('Tests app', () => {
     it('cannot GET geography metadata (wrong auth)', done => {
       request
         .get(`/geographies/${GEOGRAPHY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .expect(403)
         .end(err => {
           done(err)
@@ -805,7 +807,7 @@ describe('Tests app', () => {
     it('verifies GETing geography metadata', done => {
       request
         .get(`/geographies/${GEOGRAPHY_UUID}/meta`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(200)
         .end((err, result) => {
           test.assert(result.body.geography_metadata.some_arbitrary_thing === 'beep')
@@ -817,7 +819,7 @@ describe('Tests app', () => {
     it('verifies cannot GET non-existent geography metadata', done => {
       request
         .get(`/geographies/beepbapboop/meta`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(404)
         .end((err, result) => {
           test.assert(result.body.result === 'not found')
@@ -830,7 +832,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: 'garbage_json' }
       request
         .put(`/geographies/${GEOGRAPHY_UUID}`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .send(geography)
         .expect(403)
         .end(err => {
@@ -842,7 +844,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: 'garbage_json' }
       request
         .put(`/geographies/${GEOGRAPHY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .send(geography)
         .expect(403)
         .end(err => {
@@ -854,7 +856,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: 'garbage_json' }
       request
         .put(`/geographies/${GEOGRAPHY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(geography)
         .expect(400)
         .end((err, result) => {
@@ -867,7 +869,7 @@ describe('Tests app', () => {
       const geography = { geography_id: POLICY_UUID, geography_json: DISTRICT_SEVEN }
       request
         .put(`/geographies/${POLICY_UUID}`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(geography)
         .expect(404)
         .end((err, result) => {
@@ -880,7 +882,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: 'garbage_json' }
       request
         .post(`/geographies`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(geography)
         .expect(400)
         .end((err, result) => {
@@ -893,7 +895,7 @@ describe('Tests app', () => {
       const geography = { geography_id: GEOGRAPHY_UUID, geography_json: LA_CITY_BOUNDARY }
       request
         .post(`/geographies`)
-        .set('Authorization', SCOPED_AUTH('policies:write'))
+        .set('Authorization', POLICIES_WRITE_SCOPE)
         .send(geography)
         .expect(409)
         .end((err, result) => {
@@ -905,14 +907,14 @@ describe('Tests app', () => {
     it('cannot do bulk geography metadata reads (no auth)', async () => {
       await request
         .get(`/geographies/meta?get_read_only=false`)
-        .set('Authorization', SCOPED_AUTH())
+        .set('Authorization', EMPTY_SCOPE)
         .expect(403)
     })
 
     it('cannot do bulk geography metadata reads (wrong auth)', async () => {
       await request
         .get(`/geographies/meta?get_read_only=false`)
-        .set('Authorization', SCOPED_AUTH('events:read'))
+        .set('Authorization', EVENTS_READ_SCOPE)
         .expect(403)
     })
 
@@ -922,7 +924,7 @@ describe('Tests app', () => {
 
       const result = await request
         .get(`/geographies/meta?get_read_only=false`)
-        .set('Authorization', SCOPED_AUTH('policies:read'))
+        .set('Authorization', POLICIES_READ_SCOPE)
         .expect(200)
       test.assert(result.body.length === 2)
       test.value(result).hasHeader('content-type', APP_JSON)
