+++ packages/mds-types/index.ts
@@ -15,6 +15,8 @@
  */
 import { FeatureCollection } from 'geojson'
 
+export { AccessTokenScope, AccessTokenScopes, ScopeDescriptions, ScopeValidator } from './scopes'
+
 export const Enum = <T extends string>(...keys: T[]) =>
   Object.freeze(keys.reduce((e, key) => {
     return { ...e, [key]: key }
